src/main/webapp/app/entities/shift-manager/shift-manager.service.ts
import axios from 'axios';

import { IShiftManager } from '@/shared/model/shift-manager.model';

const baseApiUrl = 'api/shift-managers';
const baseSearchApiUrl = 'api/_search/shift-managers?query=';

export default class ShiftManagerService {
  public search(query): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(`${baseSearchApiUrl}${query}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public find(id): Promise<IShiftManager> {
    return new Promise<IShiftManager>(resolve => {
      axios.get(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public retrieve(paginationQuery?: any): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(baseApiUrl).then(function(res) {
        resolve(res);
      });
    });
  }

  public delete(id): Promise<any> {
    return new Promise<any>(resolve => {
      axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res);
      });
    });
  }

  public create(entity): Promise<IShiftManager> {
    return new Promise<IShiftManager>(resolve => {
      axios.post(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public update(entity): Promise<IShiftManager> {
    return new Promise<IShiftManager>(resolve => {
      axios.put(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }
}
